pyproject.toml
[project]
name = "callsign"
version = "1.0.0"
description = "Monitor your public IP address and notify when it changes."
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.27.2",
    "loguru>=0.7.2",
    "loguru-discord>=1.4.0",
    "python-dotenv>=1.0.1",
]

[tool.uv]
dev-dependencies = [
    "ruff>=0.6.8",
]

[tool.ruff.lint]
select = ["I"]
